+++ example/counter.ts
@@ -1,0 +1,6 @@
+import KingWorld from '../src'
+
+new KingWorld()
+	.state('counter', 0)
+	.get('/', ({ store }) => store.counter++)
+	.listen(3000)