+++ example/pure.ts
@@ -1,3 +1,6 @@
+// @ts-ignore
+console.log(caches)
+
 export default {
 	port: 8080,
 	fetch: (request: Request) => new Response('Hi')