+++ example/schema.ts
@@ -1,0 +1,17 @@
+import KingWorld from '../src'
+
+new KingWorld().post<{
+	body: {
+		username: string
+		password?: number
+	}
+	response:
+		| {
+				hello: string | number
+		  }
+		| {
+				a: 'b'
+		  }
+}>('/sign-in', () => ({
+	hello: 'world'
+}))