+++ src/context.ts
@@ -1,6 +1,10 @@
+import KingWorld from '.'
 import { TypedRoute } from './types'
 
-export default class Context<Route extends TypedRoute = TypedRoute> {
+export default class Context<
+	Route extends TypedRoute = TypedRoute,
+	Store extends KingWorld['store'] = KingWorld['store']
+> {
 	_status = 200
 	responseHeaders: Record<string, string> = {}
 
@@ -10,6 +14,7 @@ export default class Context<Route extends TypedRoute = TypedRoute> {
 		: Record<string, string>
 	params: Route['params']
 	body: Route['body']
+	store: Store
 
 	_redirect?: string
 
@@ -20,11 +25,13 @@ export default class Context<Route extends TypedRoute = TypedRoute> {
 			: Record<string, string>
 		params: Route['params']
 		body: Route['body']
+		store: Store
 	}) {
 		this.request = x.request
 		this.params = x.params
 		this.query = x.query
 		this.body = x.body
+		this.store = x.store
 	}
 
 	status = (code: number) => {