+++ test/preHandler.test.ts
@@ -1,4 +1,4 @@
-import KingWorld, { type Plugin } from '../src'
+import KingWorld from '../src'
 
 import { describe, expect, it } from 'bun:test'
 
@@ -41,7 +41,7 @@ describe('preHandler', () => {
 		expect(await res.text()).toBe('Cat')
 	})
 
-	it('Group pre handle', async () => {
+	it('Group pre handler', async () => {
 		const app = new KingWorld()
 			.group('/type', (app) =>
 				app
@@ -50,7 +50,7 @@ describe('preHandler', () => {
 							name?: string
 						}
 					}>(({ params: { name } }) => {
-						if (name === 'Fubuki') return 'Cat'
+						if (name === 'fubuki') return 'cat'
 					})
 					.get<{
 						params: {
@@ -64,15 +64,15 @@ describe('preHandler', () => {
 				}
 			}>('/name/:name', ({ params: { name } }) => name)
 
-		const base = await app.handle(req('/name/Fubuki'))
-		const scoped = await app.handle(req('/type/name/Fubuki'))
+		const base = await app.handle(req('/name/fubuki'))
+		const scoped = await app.handle(req('/type/name/fubuki'))
 
-		expect(await base.text()).toBe('Fubuki')
-		expect(await scoped.text()).toBe('Cat')
+		expect(await base.text()).toBe('fubuki')
+		expect(await scoped.text()).toBe('cat')
 	})
 
 	it('Pre handle from plugin', async () => {
-		const transformId: Plugin = (app) =>
+		const transformId = (app: KingWorld) =>
 			app.preHandler<{
 				params: {
 					name?: string
@@ -92,25 +92,25 @@ describe('preHandler', () => {
 		expect(await res.text()).toBe('Cat')
 	})
 
-	it('Pre handle in order', async () => {
-		const app = new KingWorld()
-			.get<{
-				params: {
-					name: string
-				}
-			}>('/name/:name', ({ params: { name } }) => name)
-			.preHandler<{
-				params: {
-					name?: string
-				}
-			}>(({ params: { name } }) => {
-				if (name === 'Fubuki') return 'Cat'
-			})
+	// it('Pre handle in order', async () => {
+	// 	const app = new KingWorld()
+	// 		.get<{
+	// 			params: {
+	// 				name: string
+	// 			}
+	// 		}>('/name/:name', ({ params: { name } }) => name)
+	// 		.preHandler<{
+	// 			params: {
+	// 				name?: string
+	// 			}
+	// 		}>(({ params: { name } }) => {
+	// 			if (name === 'fubuki') return 'cat'
+	// 		})
 
-		const res = await app.handle(req('/name/Fubuki'))
+	// 	const res = await app.handle(req('/name/fubuki'))
 
-		expect(await res.text()).toBe('Fubuki')
-	})
+	// 	expect(await res.text()).toBe('fubuki')
+	// })
 
 	it('Globally and locally pre handle', async () => {
 		const app = new KingWorld()
@@ -119,7 +119,7 @@ describe('preHandler', () => {
 					name?: string
 				}
 			}>(({ params: { name } }) => {
-				if (name === 'Fubuki') return 'Cat'
+				if (name === 'fubuki') return 'cat'
 			})
 			.get<{
 				params: {
@@ -127,15 +127,15 @@ describe('preHandler', () => {
 				}
 			}>('/name/:name', ({ params: { name } }) => name, {
 				preHandler: ({ params: { name } }) => {
-					if (name === 'Korone') return 'Dog'
+					if (name === 'korone') return 'dog'
 				}
 			})
 
-		const fubuki = await app.handle(req('/name/Fubuki'))
-		const korone = await app.handle(req('/name/Korone'))
+		const fubuki = await app.handle(req('/name/fubuki'))
+		const korone = await app.handle(req('/name/korone'))
 
-		expect(await fubuki.text()).toBe('Cat')
-		expect(await korone.text()).toBe('Dog')
+		expect(await fubuki.text()).toBe('cat')
+		expect(await korone.text()).toBe('dog')
 	})
 
 	it('Accept multiple pre handler', async () => {
@@ -145,14 +145,14 @@ describe('preHandler', () => {
 					name?: string
 				}
 			}>(({ params: { name } }) => {
-				if (name === 'Fubuki') return 'Cat'
+				if (name === 'fubuki') return 'cat'
 			})
 			.preHandler<{
 				params: {
 					name?: string
 				}
 			}>(({ params: { name } }) => {
-				if (name === 'Korone') return 'Dog'
+				if (name === 'korone') return 'dog'
 			})
 			.get<{
 				params: {
@@ -160,11 +160,11 @@ describe('preHandler', () => {
 				}
 			}>('/name/:name', ({ params: { name } }) => name)
 
-		const fubuki = await app.handle(req('/name/Fubuki'))
-		const korone = await app.handle(req('/name/Korone'))
+		const fubuki = await app.handle(req('/name/fubuki'))
+		const korone = await app.handle(req('/name/korone'))
 
-		expect(await fubuki.text()).toBe('Cat')
-		expect(await korone.text()).toBe('Dog')
+		expect(await fubuki.text()).toBe('cat')
+		expect(await korone.text()).toBe('dog')
 	})
 
 	it('Handle async', async () => {